create intersection between ray and plane builded from cubes position.
create simple red sphere in the scene, 3 cubes for plane and quad to visualize the plane.
using unity engine methods, create the plane using 3 cubes transform from the scene.
cast the ray from the mouse position while holding down the left mouse button,
find the intersection time on the ray when it intersects with the plane,
use the found time to get the intersection point on the ray and move the sphere to it.